py-srv/bin/requirements.txt
bottle
bottle-sqlalchemy
sqlalchemy-cockroachdb
psycopg2

# for ssl
cheroot

# for logins
beaker
requests

